deepseekv3 accurayc aligned on num_hidden_layers=4

+++ python/sgl_jax/srt/layers/attention/mla.py
@@ -77,7 +77,11 @@ class MLAAttention(nnx.Module):
                 kernel_axes=(None, None),
                 scope_name="q_a_proj",
             )
-            self.q_a_layernorm = RMSNorm(q_lora_rank, param_dtype=jnp.float32)
+            # param_dtype=fp32 keeps norm stats stable; dtype=<compute dtype>
+            # casts the output back so downstream QuantizedLinear (which
+            # preserves input dtype) returns bf16 instead of fp32, matching
+            # LinearBase's preferred_element_type=params_dtype behavior.
+            self.q_a_layernorm = RMSNorm(q_lora_rank, param_dtype=jnp.float32, dtype=dtype)
             self.q_b_proj = LinearBase(
                 q_lora_rank,
                 num_heads * self.qk_head_dim,
@@ -97,7 +101,7 @@ class MLAAttention(nnx.Module):
             kernel_axes=(None, None),
             scope_name="kv_a_proj",
         )
-        self.kv_a_layernorm = RMSNorm(kv_lora_rank, param_dtype=jnp.float32)
+        self.kv_a_layernorm = RMSNorm(kv_lora_rank, param_dtype=jnp.float32, dtype=dtype)
         self.kv_b_proj = LinearBase(
             kv_lora_rank,
             num_heads * (qk_nope_head_dim + v_head_dim),
